# Conflicts:
#	build.gradle
#	src/forge/java/team/creative/creativecore/CreativeCore.java
#	src/forge/java/team/creative/creativecore/CreativeForgeLoader.java
#	src/forge/java/team/creative/creativecore/common/config/event/ConfigEventHandler.java
#	src/main/java/team/creative/creativecore/client/render/GuiRenderHelper.java
#	src/main/java/team/creative/creativecore/client/render/text/CompiledText.java
#	src/main/java/team/creative/creativecore/client/render/text/WidthLimitedCharSink.java
#	src/main/java/team/creative/creativecore/common/config/gui/ClientSyncGuiLayer.java
#	src/main/java/team/creative/creativecore/common/config/gui/ConfigGuiLayer.java
#	src/main/java/team/creative/creativecore/common/config/gui/PlayerSelectorDialog.java
#	src/main/java/team/creative/creativecore/common/gui/GuiControl.java
#	src/main/java/team/creative/creativecore/common/gui/controls/simple/GuiButtonIcon.java
#	src/main/java/team/creative/creativecore/common/gui/controls/simple/GuiProgressbar.java
#	src/main/java/team/creative/creativecore/common/gui/controls/simple/GuiStateButton.java
#	src/main/java/team/creative/creativecore/common/gui/controls/simple/GuiTextfield.java
#	src/main/java/team/creative/creativecore/common/gui/controls/timeline/GuiTimeline.java
#	src/main/java/team/creative/creativecore/common/gui/controls/tree/GuiTreeItem.java
#	src/main/java/team/creative/creativecore/common/network/type/NetworkFieldTypes.java
#	src/main/java/team/creative/creativecore/common/util/ingredient/CreativeIngredient.java
#	src/main/java/team/creative/creativecore/common/util/ingredient/CreativeIngredientBlock.java
#	src/main/java/team/creative/creativecore/common/util/ingredient/CreativeIngredientBlockTag.java
#	src/main/java/team/creative/creativecore/common/util/ingredient/CreativeIngredientItem.java
#	src/main/java/team/creative/creativecore/common/util/ingredient/CreativeIngredientItemStack.java
#	src/main/java/team/creative/creativecore/common/util/ingredient/CreativeIngredientItemTag.java
#	src/main/java/team/creative/creativecore/common/util/text/AdvancedComponentHelper.java
#	src/main/java/team/creative/creativecore/common/util/text/TextBuilder.java
#	src/main/java/team/creative/creativecore/common/util/text/content/AdvancedFormattedText.java
#	src/main/java/team/creative/creativecore/common/util/text/content/ContentItemStack.java

+++ src/main/java/team/creative/creativecore/common/gui/manager/GuiManagerItem.java
@@ -203,7 +203,7 @@ public class GuiManagerItem extends GuiManager {
         if (rightClick)
             return 1;
         
-        return Math.max(1, Mth.floor(hand.getCount() / dragged.size()));
+        return Math.max(1, Mth.floor((float) hand.getCount() / dragged.size()));
     }
     
 }
